[FIX]: 예외처리 안되는 부분 추가
