+++ project/stampee/src/main/java/config/DBConnectionUtil.java
@@ -13,7 +13,6 @@ public class DBConnectionUtil {
 
 	public static Connection getConnection() {
 		try {
-			//라이브러리에 있는 driver를 알아서 찾음
 			Connection connection = DriverManager.getConnection(URL, USERNAME, PASSWORD);   //데이터베이스에 연결하기 위해서
 			return connection;
 		} catch (SQLException e) {
