+++ project/stampee/src/main/java/controller/StampController.java
@@ -6,8 +6,6 @@ import java.io.IOException;
 import java.net.URL;
 import java.util.List;
 import java.util.ResourceBundle;
-
-import domain.ReviewType;
 import dto.response.MyStampDto;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
